Adding custom basecommand to generate fixtures, rename files, adding custom fixtures dirs to settings.py
