+++ app/app/settings.py
@@ -64,6 +64,8 @@ TEMPLATES = [
     },
 ]
 
+FIXTURE_DIRS = [os.path.join(BASE_DIR, 'fixtures')]
+
 WSGI_APPLICATION = 'app.wsgi.application'
 
 DATABASES = {
