+++ app/static/css/style.css
@@ -1023,6 +1023,4 @@ body {
 .leaderboards_paginator_prev {
   margin-right: 10px;
 }
-
-
 /* end leaderboards-css */
